+++ src/app.module.ts
@@ -4,10 +4,12 @@ import { ProductsModule } from './products/products.module';
 import { AichatModule } from './aichat/aichat.module';
 import { UploadModule } from './upload/upload.module';
 import { PrismaModule } from './prisma/prisma.module';
+import { LoggerModule } from './logger/logger.module';
 
 @Module({
   imports: [
     ConfigModule.forRoot({ isGlobal: true }),
+    LoggerModule,
     PrismaModule,
     ProductsModule,
     AichatModule,